+++ gradle.properties
@@ -11,7 +11,7 @@ modId = taf
 
 # Root package of the mod, used to find various classes in other properties,
 # mcmod.info substitution, enabling assertions in run tasks, etc.
-modGroup = com.warpradar
+modGroup = com.taf
 
 # Whether to use modGroup as the maven publishing group.
 # Due to a history of using JitPack, the default is com.github.GTNewHorizons for all mods.
@@ -50,7 +50,7 @@ enableGenericInjection = true
 # Generate a class with a String field for the mod version named as defined below.
 # If generateGradleTokenClass is empty or not missing, no such class will be generated.
 # If gradleTokenVersion is empty or missing, the field will not be present in the class.
-generateGradleTokenClass = com.warpradar.Tags
+generateGradleTokenClass = com.taf.Tags
 
 # Name of the token containing the project's current version to generate/replace.
 gradleTokenVersion = VERSION
