+++ src/main/java/com/taf/blocks/machines/BlockServer.java
@@ -1,0 +1,5 @@
+package com.taf.blocks.machines;
+
+public class BlockServer {
+
+}
