+++ src/main/java/com/taf/main/ClientProxy.java
@@ -1,0 +1,31 @@
+package com.taf.main;
+
+import cpw.mods.fml.common.event.FMLInitializationEvent;
+import cpw.mods.fml.common.event.FMLPostInitializationEvent;
+import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+
+/**
+ * ClientProxy
+ *
+ * Обрабатывает события просходящие на стороне клиента
+ * Сами события аналогичны {@link TaF}
+ *
+ */
+
+public class ClientProxy extends CommonProxy {
+
+    @Override
+    public void preInit(FMLPreInitializationEvent event) {
+        super.preInit(event);
+    }
+
+    @Override
+    public void init(FMLInitializationEvent event) {
+        super.init(event);
+    }
+
+    @Override
+    public void postInit(FMLPostInitializationEvent event) {
+        super.postInit(event);
+    }
+}
