+++ src/main/java/com/taf/misc/satellites/luchik.java
@@ -1,0 +1,36 @@
+package com.taf.misc.satellites;
+
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.world.World;
+
+import com.hbm.entity.logic.EntityNukeExplosionMK5;
+import com.hbm.saveddata.satellites.Satellite;
+
+public class luchik extends Satellite {
+
+    public long lastOp = 0L;
+
+    public luchik() {
+        this.satIface = Interfaces.SAT_COORD;
+    }
+
+    @Override
+    public void onCoordAction(World world, EntityPlayer player, int x, int y, int z) {
+        if (lastOp + (10 * 60 * 1000) < System.currentTimeMillis()) {
+            world.spawnEntityInWorld(EntityNukeExplosionMK5.statFacNoRad(world, 300, x, y, z));
+            lastOp = System.currentTimeMillis();
+        }
+
+    }
+
+    @Override
+    public void writeToNBT(NBTTagCompound nbt) {
+        nbt.setLong("lastOp", lastOp);
+    }
+
+    @Override
+    public void readFromNBT(NBTTagCompound nbt) {
+        lastOp = nbt.getLong("lastOp");
+    }
+}
