+++ src/main/java/com/taf/registry/ModItems.java
@@ -1,0 +1,39 @@
+package com.taf.registry;
+
+import net.minecraft.item.Item;
+
+import com.hbm.items.machine.ItemSatChip;
+import com.taf.items.food.ItemEnergy;
+import com.taf.main.TaF;
+
+import cpw.mods.fml.common.registry.GameRegistry;
+
+public class ModItems {
+
+    public static Item bottle_leninade;
+    public static Item sat_luchik;
+    public static Item sat_cumone;
+
+    public static void initializeItems() {
+        bottle_leninade = new ItemEnergy().makeBottle(com.hbm.items.ModItems.bottle_empty, null)
+            .setUnlocalizedName("bottle_leninade")
+            .setTextureName(TaF.MODID + ":bottle_leninade")
+            .setCreativeTab(MainRegistry.ItemsTab);
+        sat_luchik = new ItemSatChip().setUnlocalizedName("sat_luchik")
+            .setMaxStackSize(1)
+            .setTextureName(TaF.MODID + ":sat_luchik")
+            .setCreativeTab(MainRegistry.ItemsTab);
+        sat_cumone = new ItemSatChip().setUnlocalizedName("sat_cumone")
+            .setMaxStackSize(1)
+            .setTextureName(TaF.MODID + ":sat_cumone")
+            .setCreativeTab(MainRegistry.ItemsTab);
+    }
+
+    public static void registerItems() {
+        initializeItems();
+        GameRegistry.registerItem(bottle_leninade, bottle_leninade.getUnlocalizedName());
+        GameRegistry.registerItem(sat_luchik, sat_luchik.getUnlocalizedName());
+        GameRegistry.registerItem(sat_cumone, sat_cumone.getUnlocalizedName());
+    }
+
+}
