+++ src/main/java/com/taf/registry/TileMappings.java
@@ -1,0 +1,20 @@
+package com.taf.registry;
+
+import java.util.HashMap;
+
+import net.minecraft.tileentity.TileEntity;
+
+import com.taf.tileentities.machines.TileEntityConverter;
+
+public class TileMappings {
+
+    public static HashMap<Class<? extends TileEntity>, String[]> map = new HashMap();
+
+    public static void writeMappings() {
+        put(TileEntityConverter.class, "tileentity_converter");
+    }
+
+    private static void put(Class<? extends TileEntity> cl, String... names) {
+        map.put(cl, names);
+    }
+}
