--- src/main/java/com/warpradar/blocks/machines/BlockServer.java
@@ -1,5 +1,0 @@
-package com.warpradar.blocks.machines;
-
-public class BlockServer {
-
-}
