--- src/main/java/com/warpradar/misc/utils/RuslaDoubleMath.java
@@ -1,10 +1,0 @@
-package com.warpradar.misc.utils;
-
-public class RuslaDoubleMath {
-
-    public static double nthRoot(double a, double b) {
-        double c = Math.pow(a, 1.0 / b);
-        return Math.pow(Math.round(c), b);
-    }
-
-}
