--- src/main/java/com/warpradar/registry/ModBlocks.java
@@ -1,36 +1,0 @@
-package com.warpradar.registry;
-
-import net.minecraft.block.Block;
-import net.minecraft.block.material.Material;
-
-import com.hbm.blocks.generic.BlockGeneric;
-import com.warpradar.blocks.machines.BlockConverter;
-import com.warpradar.blocks.misc.MetastableUraniumBlock;
-import com.warpradar.main.WarpRadar;
-
-import cpw.mods.fml.common.registry.GameRegistry;
-
-public class ModBlocks {
-
-    public static Block fluid_converter;
-    public static Block concrete_brick_legacy;
-    public static Block block_u238m2;
-
-    public static void initializeBlocks() {
-        fluid_converter = new BlockConverter(Material.iron).setBlockName("fluid_converter")
-            .setBlockTextureName(WarpRadar.MODID + ":fluid_converter")
-            .setCreativeTab(MainRegistry.BlocksTab);
-        concrete_brick_legacy = new BlockGeneric(Material.iron).setBlockName("concrete_brick_legacy")
-            .setBlockTextureName(WarpRadar.MODID + ":concrete_brick_legacy")
-            .setCreativeTab(MainRegistry.BlocksTab);
-        block_u238m2 = new MetastableUraniumBlock().setBlockName("block_u238m2").setBlockTextureName(WarpRadar.MODID + ":block_u238m2");
-
-    }
-
-    public static void registerBlocks() {
-        initializeBlocks();
-        GameRegistry.registerBlock(fluid_converter, fluid_converter.getUnlocalizedName());
-        GameRegistry.registerBlock(concrete_brick_legacy, concrete_brick_legacy.getUnlocalizedName());
-        GameRegistry.registerBlock(block_u238m2, block_u238m2.getUnlocalizedName());
-    }
-}
