+++ src/main/resources/mcmod.info
@@ -7,7 +7,7 @@
   "mcversion": "1.7.10",
   "url": "https://github.com/Four-el-Ltd/Trade-And-Fraction",
   "updateUrl": "https://github.com/Four-el-Ltd/Trade-And-Fraction/releases",
-  "credits": "HbMinecraft, Ruslan (coder), Kir2100 (artist), MrKimkimora (artist, coder), Vitya2127 (coder, atrist, contributor",
+  "credits": "HbMinecraft, Ruslan (coder), Kir2100 (artist), MrKimkimora (artist, coder), Vitya2127 (coder, atrist, contributor)",
   "logoFile": "assets/taf/textures/logo.png",
   "screenshots": [],
   "dependencies": []
